moved SiteTemplate to the shared directory
